Semantically irrelevant code beautification. Added missing deprecation annotations.

+++ org.semweb4j.rdfreactor.runtime/src/main/java/org/ontoware/rdfreactor/runtime/converter/ByteConverter.java
@@ -49,7 +49,7 @@ public class ByteConverter implements INodeConverter<Byte> {
 	}
 
 	public Node toNode(Model model, Object javaValue) {
-		return model.createDatatypeLiteral( ""+(Byte) javaValue, XSD._byte);
+		return model.createDatatypeLiteral( ""+javaValue, XSD._byte);
 	}
 
 }
